Move conversion functions into a web submodule
